% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_day.R
\name{pivot_day}
\alias{pivot_day}
\title{pivot_day}
\usage{
pivot_day(df)
}
\arguments{
\item{df}{A \code{data.frame} with 4 columns: day | condition | dead | censored.
\code{vial} column is optional.}
}
\value{
A \code{data.frame} in a PRISM-compatible format. First column is day,
subsequent columns are the other conditions. Each column has a series of 1(dead) and 0 (censored)
corresponding to the number of dead and censored per day.
}
\description{
Converts a "one day, multiple observations" dataframe to a PRISM-compatible
format. However, this function does not filter for days - input must all be
\strong{the same day}.
}
