% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prism.R
\name{pivot_prism}
\alias{pivot_prism}
\title{pivot_prism}
\usage{
pivot_prism(df)
}
\arguments{
\item{df}{A \code{data.frame} object with 4 columns. Column headers
must be ("condition", "day", "dead", "censored").}
}
\description{
Converts bulk survival data (multiple observations per row) to
individual survival data (one observation per row). Transforms this to a
GraphPad PRISM compatible format, with one condition per column. Empty values
are filled with NA.
}
\details{
Status: 1 = dead, 0 = censored.
}
\examples{
df_prism <- pivot_prism(sample_data)

}
