% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_surv.R
\name{plot_surv}
\alias{plot_surv}
\title{plot_surv}
\usage{
plot_surv(
  fit,
  type,
  legend.title = "Condition",
  legend.position = "right",
  xlab = "Day",
  add.median.survival = FALSE,
  add.conf.int = FALSE,
  add.pval = FALSE,
  p_adjust_method = "BH",
  ...
)
}
\arguments{
\item{fit}{a \code{survfit} object, created by \code{\link[=fit_surv]{fit_surv()}}}

\item{type}{character, either "survival" (a survival curve) or "mortality" (a mortality curve)}

\item{legend.title}{character, the legend title. For no title, use
legend.title = "". Default: "Condition"}

\item{legend.position}{character, one of type "right", "left", "bottom", or
"top". For in-plot legends, can be coordinates. c(0,0): bottom left, c(1,1): top right.}

\item{xlab}{character, the x-axis label. Default: "Day"}

\item{add.median.survival}{logical, whether to add the median survival line. Default: FALSE}

\item{add.conf.int}{logical, whether to add confidence intervals. Default: FALSE}

\item{add.pval}{logical, whether to add the p-value. Default: FALSE}

\item{p_adjust_method}{character, either "holm", "hochberg", "hommel", "bonferroni",
"BH", "BY", "fdr", "none". Default: "BH". For details, see \code{?stats::p.adjust}.}

\item{...}{additional arguments passed to \code{survminer::ggsurvplot()}.
See full list of arguments at \code{\link[survminer:ggsurvplot]{survminer::ggsurvplot()}}.}
}
\value{
A \code{ggplot2} object plotted by the \code{survminer} package
}
\description{
Plots a survival curve from \code{survfit} object.
}
\examples{
# Convert bulk survival to individual survival
df_isurv <- get_indiv_surv(sample_data,
                           sample_order = c("WT", "Drug1", "Drug2"))

# Fit survival object
surv_fit <- fit_surv(df_isurv)

# Plot survival curve
plot_surv(fit = surv_fit, type = "survival")

}
