% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bulksurv.R
\name{run_bulksurv}
\alias{run_bulksurv}
\title{run_bulksurv}
\usage{
run_bulksurv(
  sample_data,
  sample_order,
  type = "survival",
  print_stats = TRUE,
  returnData = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{sample_data}{\code{data.frame}, bulk survival data}

\item{sample_order}{character vector of conditions. Eg: c("WT", "Drug1", "Drug2")}

\item{type}{character, either "survival" (survival curve) or "mortality" (mortality curve)}

\item{print_stats}{logical, whether to print median survival, log-rank
test and pairwise log-rank test with p-value corrections. Default: TRUE}

\item{returnData}{logical, whether to return plot and statistics as a list? Default: FALSE}

\item{verbose}{logical, display messages on progress? Default: FALSE}

\item{...}{additional plot parameters passed to \code{survminer::ggsurvplot}.
Some useful parameters: \code{conf.int = TRUE}, \code{pval = TRUE}.}
}
\value{
A \code{ggplot2} object for the survival curve
}
\description{
This function performs a default survival analysis by running these steps:
\enumerate{
\item{Calculate individual survival: \code{\link{get_indiv_surv}()}}
\item{Fit survival curve by condition: \code{\link{fit_surv}()}}
\item{Plot survival curve: \code{\link{plot_surv}()}}
}
}
\examples{
#p <- run_bulksurv(sample_data,
#                    sample_order = c("WT", "Drug1", "Drug2"),
#                    type = "survival")


}
