% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_vials.R
\name{sum_vials}
\alias{sum_vials}
\title{sum_vials}
\usage{
sum_vials(df)
}
\arguments{
\item{df}{A \code{data.frame} with 4 columns: day | condition | dead | censored.
\code{vial} column is optional.}
}
\value{
A \code{data.frame}. All dead and censored values are summed across vials
for each (day, condition) pair.
}
\description{
Sums the total dead and censored across vials for each (day, condition) pair
}
\examples{

df <- tibble::tribble(
  ~day, ~condition, ~dead, ~censored, ~ vial,
  10,   "WT",       1,     0,         1,
  10,   "WT",       2,     0,         2,
  10,   "Drug1",    1,     0,         1,
  10,   "Drug1",    2,     0,         2,
  10,   "Drug2",    5,     0,         1,
  10,   "Drug1",    5,     0,         2,
  12,   "WT",       5,     0,         1,
  12,   "WT",       1,     0,         2,
  )



}
