% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_stats.R
\name{summary_stats}
\alias{summary_stats}
\title{summary_stats}
\usage{
summary_stats(df_isurv, type = "all")
}
\arguments{
\item{df_isurv}{a \code{data.frame} of survival data, with one individual per row.
If starting with bulk survival data (multiple observations per row), run \code{\link[=get_indiv_surv]{get_indiv_surv()}} first.}

\item{type}{character, either "median_survival", "logrank", "pairwise" or "all". Default: "all"}
}
\value{
If \code{type == "all"}: a \code{list} object containing results from statistical tests.
Otherwise, returns a \code{ggplot} object containing the survival/mortality curve
}
\description{
This function provides the following summary statistics:
\enumerate{
\item{\code{median_survival}: Median survival with 95\% confidence intervals}
\item{\code{logrank}: Log-rank test - is there a difference between two (or more) survival curves?}
\item{\code{pairwise}: Pairwise log-rank test with p-value correction between conditions.
Default correction: "BH". }
}
}
\details{
\itemize{
\item \code{logrank}: Calls \code{\link[survival:survdiff]{survival::survdiff()}} to test if there is a difference
between two or more survival curves using the G^rho family of tests
\item \code{pairwise}: Calls \code{\link[survminer:pairwise_survdiff]{survminer::pairwise_survdiff()}}, which calculates pairwise comparisons
between groups with corrections for multiple testing.
}
}
\examples{
df_isurv <- get_indiv_surv(sample_data,
                           sample_order = c("WT", "Drug1", "Drug2"))

stats <- summary_stats(df_isurv)
stats

}
