% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pivot_prism.R
\name{pivot_prism}
\alias{pivot_prism}
\title{pivot_prism}
\usage{
pivot_prism(sample_data, sample_order)
}
\arguments{
\item{sample_data}{A \code{data.frame} object with 4 columns. Column headers
must be ("condition", "day", "dead", "censored").}

\item{sample_order}{\code{character}, order of conditions. Eg: c("WT", "Drug1", "Drug2")}
}
\value{
A \code{tibble} arranged by \code{day}, with one condition per column. For
\code{sample_data} this is \code{day}, \code{WT}, \code{Drug1} and \code{Drug2}
}
\description{
Converts bulk survival data (multiple observations per row) to
individual survival data (one observation per row). Transforms this to a
GraphPad PRISM compatible format, with one condition per column. Empty values
are filled with NA.
}
\details{
Status: 1 = dead, 0 = censored.
}
\examples{
df_prism <- pivot_prism(sample_data,
            sample_order = c("WT", "Drug1", "Drug2"))

}
